backend/requirements.txt
# FastAPI & server
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=23.0.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
pyodbc>=5.0.0

# Settings & validation
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Azure Blob Storage
azure-storage-blob>=12.19.0
